+++ src/main/java/org/folio/entlinks/controller/converter/DataStatsMapper.java
@@ -4,7 +4,6 @@ import static org.folio.entlinks.utils.DateUtils.fromTimestamp;
 
 import java.sql.Timestamp;
 import java.time.OffsetDateTime;
-import java.util.List;
 import org.folio.entlinks.domain.dto.AuthorityStatsDto;
 import org.folio.entlinks.domain.dto.BibStatsDto;
 import org.folio.entlinks.domain.entity.AuthorityDataStat;
@@ -15,7 +14,7 @@ import org.mapstruct.Mapping;
 @Mapper(componentModel = "spring")
 public interface DataStatsMapper {
 
-  @Mapping(target = "authorityId", source = "source.authorityData.id")
+  @Mapping(target = "authorityId", source = "authorityData.id")
   @Mapping(target = "naturalIdOld", source = "authorityNaturalIdOld")
   @Mapping(target = "naturalIdNew", source = "authorityNaturalIdNew")
   @Mapping(target = "sourceFileOld", source = "authoritySourceFileOld")
@@ -24,12 +23,6 @@ public interface DataStatsMapper {
 
   @Mapping(target = "authorityNaturalId", source = "authorityData.naturalId")
   BibStatsDto convertToDto(InstanceAuthorityLink source);
-
-  default List<BibStatsDto> convertToDto(List<InstanceAuthorityLink> source) {
-    return source.stream()
-      .map(this::convertToDto)
-      .toList();
-  }
 
   default OffsetDateTime map(Timestamp timestamp) {
     return fromTimestamp(timestamp);