+++ NEWS.md
@@ -1,7 +1,7 @@
 ## v1.0.0 Released 2023-02-21
 ### APIs versions
 * Provides `instance-authority-links v1.0`
-* Provides `instance-authority-links-statistics v2.0`
+* Provides `instance-authority-links-statistics v1.0`
 * Provides `instance-authority-linking-rules v1.0`
 * Requires `login v7.0`
 * Requires `permissions v5.3`